% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLCZparameters.R
\name{getLCZparameters}
\alias{getLCZparameters}
\title{Get LCZ parameters}
\usage{
getLCZparameters(lcz_map, iStack = TRUE)
}
\arguments{
\item{lcz_map}{lcz map .tiff format}

\item{iStack}{all parameters in raster stack}
}
\value{
34 LCZ-related parameters at raster.tiff format
}
\description{
Get LCZ parameters
}
\examples{
my_par <- getLCZparametes(lcz_map)
}
