% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_cal_area.R
\name{lcz_cal_area}
\alias{lcz_cal_area}
\title{Calculate LCZ areas}
\usage{
lcz_cal_area(
  x,
  plot_type = "bar",
  iplot = TRUE,
  isave = FALSE,
  save_extension = "png",
  show_legend = TRUE,
  inclusive = FALSE,
  xlab = "LCZ code",
  ylab = "Area [km2]",
  ...
)
}
\arguments{
\item{x}{A raster SpatRaster layer containing LCZ classes.}

\item{plot_type}{Type of plot to generate: "bar", "pie", or "donut". Default is "bar".}

\item{iplot}{Logical, indicating whether to create a plot (default is TRUE).}

\item{isave}{Logical, whether to save the plot to your directory (default is FALSE).}

\item{save_extension}{File format for saving the plot. Options: "png", "jpg", "jpeg", "tif", "pdf", "svg" (default is "png").}

\item{show_legend}{Logical. If TRUE, displays the legend on the plot. If FALSE, the legend will be hidden. Default is TRUE.}

\item{inclusive}{Logical. If TRUE, use a colorblind-friendly palette. Default is FALSE.}

\item{xlab}{Character, label for the x-axis (default is "LCZ code").}

\item{ylab}{Character, label for the y-axis (default is "Area (km2)").}

\item{...}{Additional arguments to modify axis, legend, and plot labels, including title, subtitle, and caption.}
}
\value{
A summary table of LCZ class areas if plot = FALSE, otherwise, the specified plot in ggplot format.
}
\description{
This function calculates the areas of LCZ classes in both percentage and square kilometers.
}
\examples{
\dontrun{
# Calculate LCZ area and visualize as bar plot
lcz_cal_area(my_lcz_map, plot_type = "bar")

# Visualize as pie chart
lcz_cal_area(my_lcz_map, plot_type = "pie")

# Save donut chart plot
lcz_cal_area(my_lcz_map, plot_type = "donut", isave = TRUE)

# Return a dataframe
lcz_cal_area(my_lcz_map, plot_type = "donut", iplot = FALSE)
}
}
\seealso{
See the documentation for lcz_get_map() to obtain an LCZ map.
}
