% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcz_data}
\alias{lcz_data}
\title{Air Temperature LCZ Data}
\format{
A data frame with 386,354 rows and 5 columns:
\describe{
\item{date}{Date and time in POSIXct format: "2019-01-01 00:00:00"}
\item{station}{Name of the meteorological station}
\item{airT}{Air temperature in degree celsius}
\item{Latitude}{Latitude of the station}
\item{Longitude}{Longitude of the station}
}
}
\source{
\href{https://uco.berlin/}{UCO Berlin}
}
\usage{
lcz_data
}
\description{
This dataset is a subset of data from the Urban Climate Observatory (UCO) in Berlin, Germany. It contains hourly air temperature readings from 23 meteorological stations for the period of 2019-2020.
}
\examples{
data(lcz_data)
}
\keyword{datasets}
