% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_get_map2.R
\name{lcz_get_map2}
\alias{lcz_get_map2}
\title{Obtain the LCZ map (with downloaded global LCZ map)}
\usage{
lcz_get_map2(x, city = NULL, roi = NULL, isave_map = FALSE)
}
\arguments{
\item{x}{A SpatRaster object containing the LCZ map. It can be download using this link: https://zenodo.org/records/8419340/files/lcz_filter_v3.tif?download=1".
Then, it suggested use terra::rast(path/lcz_filter_v3.tif") to import the raster .tif into R.}

\item{city}{A character string specifying the name of your target area based on the OpenStreetMap project.}

\item{roi}{Optionally, you can provide a Region of Interest (ROI) in ESRI shapefile format to clip the LCZ map to a custom area.}

\item{isave_map}{Logical. Set to TRUE if you wish to save the resulting clipped map as a raster TIFF file on your local machine.}
}
\value{
A Terra raster TIFF file containing LCZ classes for the specified area of interest.
}
\description{
This function retrieves the Local Climate Zone (LCZ) global mapping dataset
created by Demuzere et al. (2022) and available at https://doi.org/10.5194/essd-14-3835-2022.
It allows you to obtain the LCZ map for a specific area of interest, which can be a city, state, region, or custom-defined shape.
}
\examples{

# Load the LCZ map from you pc
# my_map <- terra::rast(path/lcz_filter_v3.tif")

# Example 2: Get LCZ map for a custom region of interest
# custom_roi <- sf::st_read("custom_roi.shp")
# roi_lcz <- lcz_get_map2(my_map, roi = custom_roi)

# Example 3: Retrieve the LCZ map for a country (no custom ROI specified)
# my_lcz_country <- lcz_get_map2(my_map, city = "Brazil")
}
