% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_get_map_euro.R
\name{lcz_get_map_euro}
\alias{lcz_get_map_euro}
\title{Download LCZ map from the European}
\usage{
lcz_get_map_euro(
  city = NULL,
  roi = NULL,
  isave_map = FALSE,
  isave_euro = FALSE
)
}
\arguments{
\item{city}{A character string specifying the name of your target area based on the OpenStreetMap project.}

\item{roi}{Optionally, you can provide a Region of Interest (ROI) in ESRI shapefile format to clip the LCZ map to a custom area.}

\item{isave_map}{Logical. Set to TRUE if you wish to save the resulting clipped map as a raster TIFF file on your local machine.}

\item{isave_euro}{Logical. Set to TRUE if you wish to save the European LCZ map as a raster TIFF file on your local machine.}
}
\value{
A Terra raster TIFF file containing LCZ classes (1-17).
}
\description{
This function retrieves the Local Climate Zone (LCZ) European mapping dataset
created by Demuzere et al. (2019) and available at https://doi.org/10.1371/journal.pone.0214474.
It allows you to obtain the LCZ map for a specific area of interest, which can be a city, state, region, or custom-defined shape.
}
\examples{
\dontrun{
# Example 1: Load the LCZ map for a city
my_lcz_city <- lcz_get_map_euro(city = "Berlin")

# Example 2: Get LCZ map for a custom region of interest
custom_roi <- sf::st_read("custom_roi.shp")
roi_lcz <- lcz_get_map_euro(roi = custom_roi)
}
}
\references{
Demuzere, M., Bechtel, B., Middel, A., & Mills, G. (2019). Mapping Europe into local climate zones. PLOS ONE, 14(4), e0214474. DOI:https://doi.org/10.1371/journal.pone.0214474
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
