% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_get_map_generator.R
\name{lcz_get_map_generator}
\alias{lcz_get_map_generator}
\title{Download the LCZ map from LCZ Generator}
\usage{
lcz_get_map_generator(
  ID = "3110e623fbe4e73b1cde55f0e9832c4f5640ac21",
  band = "lczFilter",
  isave_map = FALSE
)
}
\arguments{
\item{ID}{A code specifying the ID generated by LCZ Factsheet (e.g, "3110e623fbe4e73b1cde55f0e9832c4f5640ac21"). More details: \url{https://lcz-generator.rub.de/factsheets/3110e623fbe4e73b1cde55f0e9832c4f5640ac21/3110e623fbe4e73b1cde55f0e9832c4f5640ac21_factsheet.html}}

\item{band}{Select the feature to use. The available options are: “lcz” and “lczFilter”. The default is "lczFilter".}

\item{isave_map}{Logical. Set to TRUE if you wish to save the resulting map as a raster TIFF file on your local machine.}
}
\value{
A Terra raster TIFF file containing LCZ classes (1-17).
}
\description{
This function retrieves the LCZ map from LCZ Generator platform
}
\examples{
\dontrun{
# Load the Rio de Janeiro map from LCZ Generator
my_lcz_map <- lcz_get_map_generator(ID = "3110e623fbe4e73b1cde55f0e9832c4f5640ac21")

}

}
\references{
Demuzere, M., Kittner, J., Bechtel, B. (2021). LCZ Generator: a web application to create Local Climate Zone maps. Frontiers in Environmental Science 9:637455. https://doi.org/10.3389/fenvs.2021.637455
}
\keyword{Climate}
\keyword{Generator}
\keyword{LCZ}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{climate,}
\keyword{urban}
