% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_get_parameters.R
\name{lcz_get_parameters}
\alias{lcz_get_parameters}
\title{Get LCZ Parameters from LCZ Stewart and Oke (2012)}
\usage{
lcz_get_parameters(x, iselect = "", istack = TRUE, ishp = FALSE, isave = FALSE)
}
\arguments{
\item{x}{A LCZ map in .tiff format SpatRaster object containing LCZ classes.}

\item{iselect}{Character vector. Specify one or more parameter names to retrieve specific
parameters. For example, "SVF1" to get the minimum Sky View Factor,
or c("SVF1", "z0") to select multiple parameters.}

\item{istack}{Logical. If TRUE, returns all parameters as a raster stack.
If FALSE, returns a list of individual parameter rasters.}

\item{ishp}{Logical. If TRUE, returns all parameters as an ESRI shapefile.}

\item{isave}{Set to TRUE if you want to save the your clipped resulting map as a raster.tiff or .shp file on your local machine.}
}
\value{
If 'istack' is TRUE, it returns a raster stack containing 34 LCZ-related parameters.
        If 'istack' is FALSE, it returns a list of individual parameter rasters.
        If 'ishp' is TRUE, it returns an ESRI shapefile containing all parameters.
        If 'iselect' is specified, it returns the selected parameter(s) as a raster(s).
}
\description{
This function extracts 34 Local Climate Zone (LCZ) parameters based on the classification
scheme developed by Stewart and Oke (2012). LCZs provide valuable information about urban
morphology and are useful for urban climate and environmental studies.
}
\examples{
# Load the LCZ map
#lcz_map <- raster::raster("path/to/lcz_map.tif")

# Get LCZ parameters as a raster stack
#lcz_params_stack <- lcz_get_parameters(lcz_map, istack = TRUE)

# Get LCZ parameters as a list of individual parameter rasters
#lcz_params_list <- lcz_get_parameters(lcz_map, istack = FALSE)

# Access individual parameters from the list
#specific_param <- lcz_params_list$parameter_name

# Get specific parameters as a raster or shapefile
#selected_params <- lcz_get_parameters(lcz_map, iselect = c("SVF1", "VEG1"), ishp = TRUE)

}
