% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_interp_map.R
\name{lcz_interp_map}
\alias{lcz_interp_map}
\title{Interpolate spatial and temporal data using a Local Climate Zones (LCZ)}
\usage{
lcz_interp_map(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  extract.method = "simple",
  sp.res = 100,
  tp.res = "hour",
  vg.model = "Sph",
  by = NULL,
  impute = NULL,
  isave = FALSE,
  LCZinterp = TRUE
)
}
\arguments{
\item{x}{A \code{SpatRaster} object representing the LCZ map, obtained via \code{lcz_get_map()} functions.}

\item{data_frame}{A data frame containing air temperature measurements with station metadata.
The data must include a date column in hourly or higher resolution.}

\item{var}{Character. The variable to interpolate (e.g., \code{"airT"} for air temperature).}

\item{station_id}{Character. The column name in \code{data_frame} identifying weather stations.}

\item{...}{Additional arguments passed to \code{selectByDate} from the \code{openair} package
for flexible time selection. Options include:
\itemize{
\item \strong{Year(s):} \code{year = 1998:2004} (selects 1998 to 2004) or \code{year = c(1998, 2004)}.
\item \strong{Month(s):} \code{month = 1:6} (January to June) or \code{month = c("January", "December")}.
\item \strong{Day(s):} \code{day = 1:30} (days 1 to 30) or \code{day = 15}.
\item \strong{Hour(s):} \code{hour = 0:23} (all hours) or \code{hour = 9}.
\item \strong{Start/End Date:} \code{"YYYY-mm-dd"} format (e.g., \code{"1999-02-01"}).
}}

\item{extract.method}{Character. The method used to assign LCZ classes to stations:
\itemize{
\item \strong{"simple"}: Assigns the LCZ based on the raster cell where the station is located. It often is used in low-density observational network.
\item \strong{"two.step"}: Filters out stations in heterogeneous LCZ areas (Daniel et al., 2017). It often is used in ultra and high-density observational network, especially in LCZ classes with multiple stations.
\item \strong{"bilinear"}: Uses bilinear interpolation to assign LCZ values.
}}

\item{sp.res}{Numeric. Spatial resolution in meters (default: \code{100}).}

\item{tp.res}{Character. Temporal resolution for averaging (\code{"hour"}, \code{"day"}, etc.).}

\item{vg.model}{Character. Variogram model for kriging. Default is \code{"Sph"} (spherical).
Available models:
\itemize{
\item \code{"Sph"}: Spherical model.
\item \code{"Exp"}: Exponential model.
\item \code{"Gau"}: Gaussian model.
\item \code{"Ste"}: M. Stein's parameterization.
}}

\item{by}{data frame time-serie split: \dQuote{year}, \dQuote{season}, \dQuote{seasonyear},  \dQuote{month}, \dQuote{monthyear}, \dQuote{weekday}, \dQuote{weekend},  \dQuote{site},
\dQuote{daylight}, \dQuote{dst} (daylight saving time). Note that the \strong{Daylight option may result in both daytime and nighttime hours being represented in UTC}. See argument \emph{type} in openair package: \url{https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option}#' @param impute Character. Method to impute missing values (\code{"mean"}, \code{"median"}, \code{"knn"}, \code{"bag"}).}

\item{isave}{Logical. If \code{TRUE}, saves the plot to the working directory.}

\item{LCZinterp}{Logical. If \code{TRUE} (default), applies LCZ-based interpolation.}
}
\value{
A spatially interpolated map in \code{terra raster GeoTIFF} format.
}
\description{
Generates an interpolated spatial map using LCZ as background.
This function applies geostatistical interpolation techniques (e.g., kriging) to estimate variable values (e.g., air temperature) across the study area,
integrating LCZ classifications. The function allows for flexible time period selection, like hour, daytime, nighttime and so on.
}
\examples{
\dontrun{
# Load sample data and LCZ map
data("lcz_data")
lcz_map <- lcz_get_map_generator(ID = "8576bde60bfe774e335190f2e8fdd125dd9f4299")

# Plot the LCZ map
lcz_plot_map(lcz_map)

# Perform interpolation for 2020-01-02 at 04:00h
my_interp <- lcz_interp_map(
  lcz_map, data_frame = lcz_data, var = "airT",
  station_id = "station", year = 2020, month = 1, day = 2, hour = 4,
  sp.res = 100, tp.res = "hour", vg.model = "Sph",
  LCZinterp = TRUE
)

# Perform interpolation for an entire day (returns a raster stack of 24 hours)
my_interp <- lcz_interp_map(
  lcz_map, data_frame = lcz_data, var = "airT",
  station_id = "station", year = 2020, month = 1, day = 2,
  sp.res = 100, tp.res = "hour", vg.model = "Sph",
  LCZinterp = TRUE
)
}

}
\references{
Anjos, M., Targino, A. C., Krecl, P., Oukawa, G. Y. & Braga, R. F. Analysis of the urban heat island under different synoptic patterns using local climate zones. Build. Environ. 185, 107268 (2020).
Fenner, D., Meier, F., Bechtel, B., Otto, M. & Scherer, D. (2017). Intra and inter ‘local climate zone’ variability of air temperature as observed by crowdsourced citizen weather stations in Berlin, Germany. \emph{Meteorol. Z.}, 26, 525–547.
\url{http://www.gstat.org/}
}
\seealso{
See \code{lcz_inter_eval()} for LCZ interpolation evaluation.
}
\author{
Max Anjos (\url{https://github.com/ByMaxAnjos})
}
\keyword{Climate}
\keyword{Local}
\keyword{Zone,}
\keyword{interpolation,}
\keyword{kriging.}
