% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_plot_parameters.R
\name{lcz_plot_parameters}
\alias{lcz_plot_parameters}
\title{Plot LCZ parameter map}
\usage{
lcz_plot_parameters(
  x,
  iselect = "",
  isubtitle = "",
  all = FALSE,
  isave = FALSE
)
}
\arguments{
\item{x}{The LCZ map in SpatRaster format.}

\item{iselect}{Character vector. Specify one or more parameter names to retrieve specific
parameters. For example, "SVF.min" to get the minimum Sky View Factor,
or c("SVF.min", "veg.frac.max") to select multiple parameters.}

\item{isubtitle}{Specify your name area}

\item{all}{Salve all parameters into LCZ4r_output}

\item{isave}{Save the plot into your directory.}
}
\value{
A plot of the LCZ parameters
}
\description{
This function plots the parameters of an LCZ (Local Climate Zone) map.
}
\examples{

#myplot_par <- lcz_plot_parameters(x = lcz_map, iselect = "SVF.min", isubtitle = "Your City")

}
\seealso{
See the documentation for lcz_get_parameters() to obtain an LCZ parameters
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
