% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_tutorial.R
\name{lcz_tutorial}
\alias{lcz_tutorial}
\title{LCZ4r tutorials}
\usage{
lcz_tutorial(article = "general")
}
\arguments{
\item{article}{String. Default is "general" for general functions. It can be either "local" for local functions}
}
\value{
Opens LCZ4r tutorial on a web browser
}
\description{
Opens LCZ4r tutorial on a web browser. This function requires internet connection
}
\examples{
lcz_tutorial(article="general")
}
