% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_uhi_intensity.R
\name{lcz_uhi_intensity}
\alias{lcz_uhi_intensity}
\title{Assess the Urban Heat Island Intensity}
\usage{
lcz_uhi_intensity(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  extract.method = "simple",
  method = "LCZ",
  Turban = NULL,
  Trural = NULL,
  by = NULL,
  impute = NULL,
  group = FALSE,
  iplot = TRUE,
  isave = FALSE,
  save_extension = "png",
  ylab = "Air temperature [C]",
  xlab = "Time",
  ylab2 = "UHI",
  title = "",
  caption = "LCZ4r,2024"
)
}
\arguments{
\item{x}{A \code{SpatRaster} object containing the LCZ map. The LCZ map can be obtained using the \code{lcz_get_map()} function.}

\item{data_frame}{A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.}

\item{var}{The name of the variable in the data frame representing air temperature.}

\item{station_id}{The name of the variable in the data frame representing station IDs.}

\item{...}{Additional arguments for the \code{selectByDate} function from the \code{openair} package. These arguments allow for flexible selection of specific time periods (year, month, day, hour). Examples of how to use these arguments include:
\itemize{
\item \strong{Year(s)}: Numeric value(s) specifying the year(s) to select. For example, \code{year = 1998:2004} selects all years between 1998 and 2004 (inclusive), while \code{year = c(1998, 2004)} selects only the years 1998 and 2004.
\item \strong{Month(s)}: Numeric or character value(s) specifying the months to select. Numeric examples: \code{month = 1:6} (January to June), or character examples: \code{month = c("January", "December")}.
\item \strong{Day(s)}: Numeric value(s) specifying the days to select. For instance, \code{day = 1:30} selects days from 1 to 30, or \code{day = 15} selects only the 15th day of the month.
\item \strong{Hour(s)}: Numeric value(s) specifying the hours to select. For example, \code{hour = 0:23} selects all hours in a day, while \code{hour = 9} selects only the 9th hour.
\item \strong{Start date}: A string specifying the start date in either start="DD/MM/YYYY" (e.g., "1/2/1999") or "YYYY-mm-dd" format (e.g., "1999-02-01").
\item \strong{End date}: A string specifying the start date in either end="DD/MM/YYYY" (e.g., "1/2/1999") or "YYYY-mm-dd" format (e.g., "1999-02-01").
}}

\item{time.freq}{The time period to average to. Default is \dQuote{hour}, but can also be \dQuote{day}, \dQuote{week}, \dQuote{month}, or \dQuote{year}.}

\item{extract.method}{A character string specifying the method used to assign the LCZ class to each station point. The default is \code{"simple"}. Available methods are:
\itemize{
\item \strong{simple}: Assigns the LCZ class based on the value of the raster cell in which the point falls. It often is used in low-density observational network.
\item \strong{two.step}: Assigns LCZs to stations while filtering out those located in heterogeneous LCZ areas. This method requires that at least 80\% of the pixels within a 5 × 5 kernel match the LCZ of the center pixel (Daniel et al., 2017). Note that this method reduces the number of stations. It often is used in ultra and high-density observational network, especially in LCZ classes with multiple stations.
\item \strong{bilinear}: Interpolates the LCZ class values from the four nearest raster cells surrounding the point.
}}

\item{method}{Method to calculate the UHI intensity. Options include: LCZ and manual. The LCZ method automatically identifies the LCZ build types, starting from LCZ 1 and progressing to LCZ 10, to represent the urban temperature, then it starts from LCZ natural LCZ (11-16) to represent the rural temperature. The manual method is a character string indicating the stations as references for the urban and rural areas.}

\item{Turban}{Urban station references, if the method "manual" is selected.}

\item{Trural}{Rural station references, if the method "manual" is selected.}

\item{by}{data frame time-serie split: \dQuote{year}, \dQuote{season}, \dQuote{seasonyear},  \dQuote{month}, \dQuote{monthyear}, \dQuote{weekday}, \dQuote{weekend},  \dQuote{site},
\dQuote{daylight}, \dQuote{dst} (daylight saving time). Note that the \strong{Daylight option may result in both daytime and nighttime hours being represented in UTC}. See argument \emph{type} in openair package: \url{https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option}}

\item{impute}{Method to impute missing values. Options include \dQuote{mean}, \dQuote{median}, \dQuote{knn}, or \dQuote{bag}.}

\item{group}{If TRUE urban, rural and UHI - related temperatures are grouped in the same plot.}

\item{iplot}{Set to \code{TRUE} to return a plot. If \code{FALSE}, a data frame is returned.}

\item{isave}{Set to \code{TRUE} to save all results (plot, time-series, station references) into your directory.}

\item{save_extension}{A character string indicating the file format for saving the plot. Options include: \dQuote{png}, \dQuote{jpg}, \dQuote{jpeg}, \dQuote{tif}, \dQuote{pdf}, \dQuote{svg}. The default is \dQuote{png}.}

\item{ylab}{The label for the y-axis.}

\item{xlab}{The label for the x-axis. Default is \dQuote{Time}.}

\item{ylab2}{The label for the right y-axis.}

\item{title}{The title of the plot. Default is an empty string.}

\item{caption}{The source data caption.}
}
\value{
A visual representation of the time series of UHI in \code{ggplot} or data frame .csv format.
}
\description{
This function calculates the Urban Heat Island (UHI) intensity based on air temperature measurements and Local Climate Zones (LCZ). More details: \url{https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html}
}
\examples{
\dontrun{
# Calculate hourly UHI intensity for 2019 using the LCZ method
my_uhi <- lcz_uhi_intensity(lcz_map, data_frame = lcz_data, var = "airT",
                            station_id = "station", year = 2019,
                            time.freq = "hour", method = "LCZ")


}

}
\seealso{
See the documentation for \code{lcz_get_map()} to obtain an LCZ map.
}
\author{
Max Anjos (\url{https://github.com/ByMaxAnjos})
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
