% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCZarea.R
\name{LCZarea}
\alias{LCZarea}
\title{Calculate LCZ Class Areas}
\usage{
LCZarea(lcz_map = NULL, iplot = TRUE)
}
\arguments{
\item{lcz_map}{A raster Spatraster layer containing LCZ classes.}

\item{iplot}{Logical, indicating whether to create a plot (default is TRUE).}
}
\value{
A summary table of LCZ class areas if iplot is FALSE, otherwise, a bar plot.
}
\description{
This function calculates the areas of LCZ (Local Climate Zones) classes in both percentage and square kilometers.
}
\examples{
myLCZarea <- LCZarea(map, iplot = TRUE)

}
