% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_plot_map.R
\name{lcz_plot_map}
\alias{lcz_plot_map}
\title{Plot LCZ Map}
\usage{
lcz_plot_map(x, isubtitle = "", isave = FALSE, legend = "name")
}
\arguments{
\item{x}{The LCZ map in SpatRaster format.}

\item{isubtitle}{Specify your name area}

\item{isave}{Save the plot into your directory.}

\item{legend}{Specify whether to include a legend. Set to "code" to show LCZ codes in the legend.}
}
\value{
A plot of the LCZ map.
}
\description{
This function plots the classes of an LCZ (Local Climate Zone) map represented as a SpatRaster object.
}
\examples{

#myplot <- lcz_plot_map(x = myLCZmap, legend = "name")

}
\seealso{
See the documentation for getLCZmap() to obtain an LCZ map.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
