% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLCZmap.R
\name{getLCZmap}
\alias{getLCZmap}
\title{Get your LCZ map}
\usage{
getLCZmap(city = NULL, roi = NULL, isave = TRUE)
}
\arguments{
\item{city}{name your city}

\item{roi}{or use region of interest at ESRI shapefile format}

\item{isave}{if you want to save the map.tiff into you machine}
}
\value{
a raster.tiff with the LCZ classes
}
\description{
Get your LCZ map
}
\examples{
myLczmap <- getLCZmap(city = "Berlin")
}
