% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_anomaly.R
\name{lcz_anomaly}
\alias{lcz_anomaly}
\title{Explore LCZ Thermal Anomalies}
\usage{
lcz_anomaly(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  by = NULL,
  hemisphere = "northern",
  impute = NULL,
  iplot = FALSE,
  isave = FALSE,
  inclusive = FALSE,
  ylab = "Thermal Anomaly [Degree Celsius]",
  xlab = "Station",
  title = "",
  caption = "Source: LCZ4r; Stewart and Oke, 2012; Demuzere et al.2022."
)
}
\arguments{
\item{x}{A SpatRaster object containing the LCZ map. The LCZ map can be obtained using the lcz_get_map() function.}

\item{data_frame}{A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.}

\item{var}{Name of the variable, e.g. air temperature, in the dataframe.}

\item{station_id}{Name of the station ID variable in the dataframe.}

\item{...}{Utilities from \code{selectBydata} from \code{openair} package. A start date string in the form e.g. \dQuote{1/2/1999} or in format i.e. \dQuote{YYYY-mm-dd}, \dQuote{1999-02-01}.
A year or years to select e.g. year = 1998:2004 to select 1998-2004 inclusive or year = c(1998, 2004) to select 1998 and 2004. A month or months to select.
Can either be numeric e.g. month = 1:6 to select months 1-6 (January to June), or by name e.g. month = c(\dQuote{January}, \dQuote{December}).}

\item{time.freq}{Defines the time period to average to. Default is \dQuote{hour}.}

\item{by}{Temporal components of the data (e.g.,\dQuote{sec}, \dQuote{min}, \dQuote{hour}, \dQuote{day}, \dQuote{DSTday},
\dQuote{week}, \dQuote{month}, \dQuote{quarter} or \dQuote{year}). See argument \emph{type} in openair package.}

\item{hemisphere}{Hemisphere for splitting data into seasons (\dQuote{northern} or \dQuote{southern}).}

\item{impute}{Method to impute missing values (\dQuote{mean}, \dQuote{median}, \dQuote{knn}, \dQuote{bag}).}

\item{iplot}{Set to TRUE if you want to save the plot in your working directory.}

\item{isave}{Save the plot into your directory.}

\item{inclusive}{Set to TRUE to a colorblind-friendly palette.}

\item{ylab}{y-axis name. Default is \dQuote{Thermal Anomaly Degree Celsius}.}

\item{xlab}{y-axis name. Default is \dQuote{Station}}

\item{title}{y-axis name. Default is \dQuote{" "}.}

\item{caption}{source data. Default is \dQuote{Source: LCZ4r, Stewart and Oke, 2012; Demuzere et al.2022."}.}
}
\value{
A visual representation of the time series of air temperature of LCZ in \code{ggplot} format
}
\description{
This function generates a graphical representation of thermal anomaly for different Local Climate Zones (LCZs)
}
\examples{

# Hourly air temperature values in 2019.
# my_ts <- lcz_anomaly(my_map, df = air_UCON, var = "airT", station_id = "station", year = 2019)

}
\seealso{
See the documentation for lcz_get_map() to obtain an LCZ map.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
