% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_anomaly_map.R
\name{lcz_anomaly_map}
\alias{lcz_anomaly_map}
\title{Visualize the interpolated LCZ-Thermal Anomaly map}
\usage{
lcz_anomaly_map(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  sp.res = 100,
  tp.res = "hour",
  method = "Kriging",
  ...,
  by = NULL,
  impute = NULL,
  isave = FALSE
)
}
\arguments{
\item{x}{A SpatRaster object containing the LCZ map. The LCZ map can be obtained using the lcz_get_map() function.}

\item{data_frame}{A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.}

\item{var}{Name of the variable for interpolation, e.g. air temperature, in the dataframe.}

\item{station_id}{Name of the station ID variable in the dataframe.}

\item{sp.res}{Spatial resolution in unit of meters for interpolation. Default is 100.}

\item{tp.res}{Temporal resolution, the time period to average to. Default is \dQuote{hour}, but includes \dQuote{day}, \dQuote{week}, \dQuote{month} or \dQuote{year}.}

\item{method}{Statistical technique used in the interpolation, such as kriging and IDW. Default is "kriging".}

\item{...}{Utilities from \code{selectBydata} from \code{openair} package. A start date string in the form e.g. \dQuote{1/2/1999} or in format i.e. \dQuote{YYYY-mm-dd}, \dQuote{1999-02-01}.
A year or years to select e.g. year = 1998:2004 to select 1998-2004 inclusive or year = c(1998, 2004) to select 1998 and 2004. A month or months to select.
Can either be numeric e.g. month = 1:6 to select months 1-6 (January to June), or by name e.g. month = c(\dQuote{January}, \dQuote{December}).}

\item{by}{data frame time-serie split: \dQuote{year}, \dQuote{season}, \dQuote{seasonyear},  \dQuote{month}, \dQuote{monthyear}, \dQuote{weekday}, \dQuote{weekend},  \dQuote{site},
\dQuote{daylight}, (daylight saving time).See argument \emph{type} in openair package: https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option}

\item{impute}{Method to impute missing values (\dQuote{mean}, \dQuote{median}, \dQuote{knn}, \dQuote{bag}).}

\item{isave}{Save the plot into your directory.}
}
\value{
A map of LCZ-thermal anomalies in \code{terra raster.tif} format
}
\description{
This function generates a spatial interpolation of thermal anomaly for different Local Climate Zones (LCZs). More details: https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html
}
\examples{

# Hourly air temperature values in 2019.
}
\seealso{
See the documentation for lcz_get_map() to obtain an LCZ map.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
