% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_cal_area.R
\name{lcz_cal_area}
\alias{lcz_cal_area}
\title{Calculate LCZ areas}
\usage{
lcz_cal_area(x, iplot = TRUE, isave = FALSE, inclusive = FALSE)
}
\arguments{
\item{x}{A raster SpatRaster layer containing LCZ classes.}

\item{iplot}{Logical, indicating whether to create a plot (default is TRUE).}

\item{isave}{Save the plot into your directory.}

\item{inclusive}{Set to TRUE to a colorblind-friendly palette.}
}
\value{
A summary table of LCZ class areas if iplot is FALSE, otherwise, a bar plot.
}
\description{
This function calculates the areas of LCZ classes in both percentage and square kilometers.
}
\examples{

#Calcute the LCZ area
#my_lcz_area <- lcz_cal_area(x= my_lcz_map, iplot = TRUE, isave = TRUE)

}
\seealso{
See the documentation for lcz_get_map() to obtain an LCZ map.
}
