% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcz_data}
\alias{lcz_data}
\title{Air temperature LCZ data}
\format{
A data frame with 386,354 rows and 5 columns:
\describe{
  \item{date}{date POSIXct, format: "2019-01-01 00:00:00"}
  \item{station}{station name}
  \item{Latitude}{Latitude station}
  \item{Longitue}{Longitude station}
  ...
}
}
\source{
<https://uco.berlin/>
}
\usage{
lcz_data
}
\description{
A subset of data from the Urban Climate Observatory (UCO) Berlin, Germany.
It contains hourly air temperature readings from 23 meteorological stations for the 2019-2020 period.
}
\examples{
data(lcz_data)
}
\keyword{datasets}
