% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_get_map.R
\name{lcz_get_map}
\alias{lcz_get_map}
\title{Download the LCZ map}
\usage{
lcz_get_map(city = NULL, roi = NULL, isave_map = FALSE, isave_global = FALSE)
}
\arguments{
\item{city}{A character string specifying the name of your target area based on the OpenStreetMap project.}

\item{roi}{Optionally, you can provide a Region of Interest (ROI) in ESRI shapefile format to clip the LCZ map to a custom area.}

\item{isave_map}{Logical. Set to TRUE if you wish to save the resulting clipped map as a raster TIFF file on your local machine.}

\item{isave_global}{Logical. Set to TRUE if you wish to save the global LCZ map as a raster TIFF file on your local machine.}
}
\value{
A Terra raster TIFF file containing LCZ classes (1-17).
}
\description{
This function retrieves the Local Climate Zone (LCZ) global mapping dataset
It allows you to obtain the LCZ map for a specific area of interest, which can be a city, state, region, or custom-defined shape.
}
\examples{
\dontrun{
# Load the LCZ map for a city
my_lcz_city <- lcz_get_map(city = "your city")

# Get LCZ map for a custom region of interest
custom_roi <- sf::st_read("custom_roi.shp")
roi_lcz <- lcz_get_map(roi = custom_roi, isave_map = TRUE)

# Retrieve the LCZ map for a country
my_lcz_country <- lcz_get_map(city = "Brazil")
}

}
\references{
Demuzere, M., Kittner, J., Martilli, A., Mills, G., Moede, C., Stewart, I. D., van Vliet, J., and Bechtel, B. (2022). A global map of Local Climate Zones to support earth system modelling and urban scale environmental science, Earth Syst. Sci. Data 14(8) 3835-3873. DOI:https://doi.org/10.5194/essd-14-3835-2022
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
