% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_plot_interp.R
\name{lcz_plot_interp}
\alias{lcz_plot_interp}
\title{Visualize Interpolated LCZ map}
\usage{
lcz_plot_interp(
  x,
  palette = "muted",
  direction = 1,
  ncol = NULL,
  nrow = NULL,
  isave = FALSE,
  save_extension = "png",
  ...
)
}
\arguments{
\item{x}{The LCZ map in SpatRaster format, either as a single raster or a raster stack, obtained from functions like lcz_anomaly_map() or lcz_interp_map().}

\item{palette}{Gradient palettes available in the tidyterra package. Default is "muted". More palettes can be found in the tidyterra documentation: https://dieghernan.github.io/tidyterra/articles/palettes.html#scale_fill_whitebox_}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{ncol}{Number of columns in the plot.}

\item{nrow}{Number of rows in the plot.}

\item{isave}{Logical, indicating whether to save the plot to your directory. Default is FALSE.}

\item{save_extension}{File format for saving the plot. Options: "png", "jpg", "jpeg", "tif", "pdf", "svg" (default is "png").}

\item{...}{Additional arguments to modify axis, legend, and plot labels, including title, subtitle, and caption.}
}
\value{
A plot of the LCZ interpolated map in ggplot2 format.
}
\description{
This function plots the interpolated LCZ anomaly, LCZ air temperature, or other environmental variables.
}
\examples{
\dontrun{
 # Plot the interpolated map
 lcz_plot_interp(my_interp_map)
 }

}
\seealso{
See the documentation for lcz_anomaly_map() to obtain LCZ anomaly rasters and lcz_interp_map() for air temperature rasters.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
