% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_plot_parameters.R
\name{lcz_plot_parameters}
\alias{lcz_plot_parameters}
\title{Visualize LCZ parameters}
\usage{
lcz_plot_parameters(
  x,
  iselect = "",
  all = FALSE,
  inclusive = FALSE,
  isave = FALSE,
  save_extension = "png",
  ...
)
}
\arguments{
\item{x}{The LCZ map in SpatRaster in a stack format.}

\item{iselect}{Character vector. Specify one or more parameter names to retrieve specific
parameters. For example, "SVF2" to get the minimum Sky View Factor,
or c("z0", "BSF2") to select multiple parameters. See the complete  parameter list here: \url{https://bymaxanjos.github.io/LCZ4r/articles/Introd_general_LCZ4r.html#retrieve-and-visualize-lcz-parameters}.}

\item{all}{Logical, specifying whether to save all selected parameters into LCZ4r_output. Default is FALSE.}

\item{inclusive}{Set to TRUE to a colorblind-friendly palette.}

\item{isave}{Logical, indicating whether to save the plot to your directory. Default is FALSE.}

\item{save_extension}{File format for saving the plot. Options: "png", "jpg", "jpeg", "tif", "pdf", "svg" (default is "png").}

\item{...}{An optional modify axis, legend, and plot labels: subtitle, and caption. It does not work with title.}
}
\value{
A plot of the selected LCZ parameters in ggplot2 format
}
\description{
This function plots the parameters of an Local Climate Zone map.
}
\examples{
\dontrun{
# Plot the minimum Sky View Factor (SVF1) for your city
lcz_plot_parameters(lcz_map, iselect = "SVF1")

# Plot multiple parameters and save them to the LCZ4r_output directory
lcz_plot_parameters(lcz_map, iselect = c("BSF3", "AH3"))
}
}
\seealso{
See the documentation for lcz_get_parameters() to obtain LCZ parameters.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
