% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_plot_parameters.R
\name{lcz_plot_parameters}
\alias{lcz_plot_parameters}
\title{Visualize LCZ parameters}
\usage{
lcz_plot_parameters(
  x,
  iselect = "",
  all = FALSE,
  inclusive = FALSE,
  isave = FALSE,
  save_extension = "png",
  ...
)
}
\arguments{
\item{x}{The LCZ map in SpatRaster in a stack format.}

\item{iselect}{Code vector. Specify one or more parameter:
\itemize{
\item \strong{SVFmean}: Mean Sky View Factor
\item \strong{SVFmax}: Maximum Sky View Factor
\item \strong{SVFmin}: Minimum Sky View Factor
\item \strong{ARmean}: Mean Aspect Ratio
\item \strong{ARmax}: Maximum Aspect Ratio
\item \strong{ARmin}: Minimum Aspect Ratio
\item \strong{BSFmean}: Mean Building Surface Fraction
\item \strong{BSFmax}: Maximum Building Surface Fraction
\item \strong{BSFmin}: Minimum Building Surface Fraction
\item \strong{ISFmean}: Mean Impervious Surface Fraction
\item \strong{ISFmax}: Maximum Impervious Surface Fraction
\item \strong{ISFmin}: Minimum Impervious Surface Fraction
\item \strong{PSFmean}: Mean Vegetation Surface Fraction
\item \strong{PSFmax}: Maximum Vegetation Surface Fraction
\item \strong{PSFmin}: Minimum Vegetation Surface Fraction
\item \strong{TSFmean}: Mean Tree Surface Fraction
\item \strong{TSFmax}: Maximum Tree Surface Fraction
\item \strong{TSFmin}: Minimum Tree Surface Fraction
\item \strong{HREmean}: Mean Height Roughness Elements
\item \strong{HREmax}: Maximum Height Roughness Elements
\item \strong{HREmin}: Minimum Height Roughness Elements
\item \strong{TRCmean}: Mean Terrain Roughness class
\item \strong{TRCmax}: Maximum Terrain Roughness class
\item \strong{TRCmin}: Minimum Terrain Roughness class
\item \strong{SADmean}: Mean Surface Admittance
\item \strong{SADmax}: Maximum Surface Admittance
\item \strong{SADmin}: Minimum Surface Admittance
\item \strong{SALmean}: Mean Surface Albedo
\item \strong{SALmax}: Maximum Surface Albedo
\item \strong{SALmin}: Minimum Surface Albedo
\item \strong{AHmean}: Mean Anthropogenic Heat Outupt
\item \strong{AHmax}: Maximum Anthropogenic Heat Outupt
\item \strong{AHmin}: Minimum Anthropogenic Heat Outupt
\item \strong{z0}: Roughness Lenght class
}}

\item{all}{Logical, specifying whether to save all selected parameters into LCZ4r_output. Default is FALSE.}

\item{inclusive}{Set to TRUE to a colorblind-friendly palette.}

\item{isave}{Logical, indicating whether to save the plot to your directory. Default is FALSE.}

\item{save_extension}{File format for saving the plot. Options: "png", "jpg", "jpeg", "tif", "pdf", "svg" (default is "png").}

\item{...}{An optional modify axis, legend, and plot labels: subtitle, and caption. It does not work with title.}
}
\value{
A plot of the selected LCZ parameters in ggplot2 format

#'@references
Stewart, I., and T. Oke, 2012: Local climate zones for urban temperature studies. Bull. Amer. Meteor. Soc., 93, 1879–1900. https://doi.org/10.1175/BAMS-D-11-00019.1
}
\description{
This function plots the parameters of an Local Climate Zone map.
}
\examples{
\dontrun{
# Plot the mean Sky View Factor (SVFmean) for your city
lcz_plot_parameters(lcz_map, iselect = "SVFmean")

# Plot multiple parameters and save them to the LCZ4r_output directory
lcz_plot_parameters(lcz_map, iselect = c("BSFmean", "AHmean"))
}
}
\seealso{
See the documentation for lcz_get_parameters() to obtain LCZ parameters.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
