% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_ts.R
\name{lcz_ts}
\alias{lcz_ts}
\title{Analyze LCZ Time Series}
\usage{
lcz_ts(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  by = NULL,
  impute = NULL,
  iplot = TRUE,
  isave = FALSE,
  palette = "VanGogh2",
  ylab = "Air temperature [Degree Celsius]",
  xlab = "Time",
  title = "",
  caption = ""
)
}
\arguments{
\item{x}{A \code{SpatRaster} object containing the LCZ map. The LCZ map can be obtained using the \code{lcz_get_map()} function.}

\item{data_frame}{A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.}

\item{var}{The name of the variable in the data frame representing air temperature.}

\item{station_id}{The name of the variable in the data frame representing station IDs.}

\item{...}{Additional arguments for the \code{selectBydata} from \code{openair} package. These include:
\itemize{
\item A start date string in the form "1/2/1999" or in format "YYYY-mm-dd", e.g., "1999-02-01".
\item A year or years to select, e.g., year = 1998:2004 to select 1998-2004 inclusive, or year = c(1998, 2004) to select 1998 and 2004.
\item A month or months to select. Can either be numeric, e.g., month = 1:6 to select January to June, or by name, e.g., month = c("January", "December").
}}

\item{time.freq}{Defines the time period to average to. Default is \dQuote{hour}, but includes \dQuote{day}, \dQuote{week}, \dQuote{month} or \dQuote{year}.}

\item{by}{data frame time-serie split: \dQuote{year}, \dQuote{season}, \dQuote{seasonyear},  \dQuote{month}, \dQuote{monthyear}, \dQuote{weekday}, \dQuote{weekend},  \dQuote{site},
\dQuote{daylight}, \dQuote{dst} (daylight saving time).See argument \emph{type} in openair package: \url{https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option}}

\item{impute}{Method to impute missing values (\dQuote{mean}, \dQuote{median}, \dQuote{knn}, \dQuote{bag}).}

\item{iplot}{Set to \code{TRUE} to return a plot. If \code{FALSE}, a data frame is returned.}

\item{isave}{Set to \code{TRUE} to save all results (plot, time-series) into your directory.}

\item{palette}{Default is "VanGogh2". Define your color palette from MetBrewer \url{https://github.com/BlakeRMills/MetBrewer?tab=readme-ov-file#palettes}.}

\item{ylab}{y-axis name.}

\item{xlab}{y-axis name. Default is \dQuote{Time}}

\item{title}{y-axis name. Default is \dQuote{" "}.}

\item{caption}{source data. Default can be \dQuote{Source:LCZ4r,2024"}.}
}
\value{
A visual representation of the time series of air temperature of LCZ in \code{ggplot} or data frame .csv format.
}
\description{
This function generates a graphical representation of time series air temperature data for different Local Climate Zones (LCZs). More details: \url{https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html}
}
\examples{
\dontrun{
# Hourly air temperature values in 2019.
my_ts <- lcz_ts(lcz_map, df = lcz_data, var = "airT",
                 station_id = "station", year = 2019)

}
}
\seealso{
See the documentation for  \code{lcz_get_map()} to obtain an LCZ map.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
