% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_tutorial.R
\name{lcz_tutorial}
\alias{lcz_tutorial}
\title{LCZ4r Tutorials}
\usage{
lcz_tutorial(x = "general")
}
\arguments{
\item{x}{Character string. Either "general" for general functions or "local" for local functions. Default is "general".}
}
\value{
Opens the LCZ4r tutorial in a web browser.
}
\description{
Opens LCZ4r tutorial in a web browser. This function requires an internet connection.
}
\examples{
\dontrun{
lcz_tutorial("general")
lcz_tutorial("local")
}
}
