% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz_uhi_intensity.R
\name{lcz_uhi_intensity}
\alias{lcz_uhi_intensity}
\title{Assess the Urban Heat Island Intensity}
\usage{
lcz_uhi_intensity(
  x,
  data_frame = "",
  var = "",
  station_id = "",
  ...,
  time.freq = "hour",
  method = "LCZ",
  Turban = NULL,
  Trural = NULL,
  by = NULL,
  impute = NULL,
  group = FALSE,
  iplot = TRUE,
  isave = FALSE,
  ylab = "Air temperature [C]",
  xlab = "Time",
  ylab2 = "UHI",
  title = "",
  caption = "LCZ4r,2024"
)
}
\arguments{
\item{x}{A \code{SpatRaster} object containing the LCZ map. The LCZ map can be obtained using the \code{lcz_get_map()} function.}

\item{data_frame}{A data frame containing air temperature measurements and station IDs. The data frame should have a date field in hourly or higher resolution format.}

\item{var}{The name of the variable in the data frame representing air temperature.}

\item{station_id}{The name of the variable in the data frame representing station IDs.}

\item{...}{Additional arguments for the \code{selectBydata} from \code{openair} package. These include:
\itemize{
\item A start date string in the form "1/2/1999" or in format "YYYY-mm-dd", e.g., "1999-02-01".
\item A year or years to select, e.g., year = 1998:2004 to select 1998-2004 inclusive, or year = c(1998, 2004) to select 1998 and 2004.
\item A month or months to select. Can either be numeric, e.g., month = 1:6 to select January to June, or by name, e.g., month = c("January", "December").
}}

\item{time.freq}{The time period to average to. Default is \dQuote{hour}, but can also be \dQuote{day}, \dQuote{week}, \dQuote{month}, or \dQuote{year}.}

\item{method}{Method to calculate the UHI intensity. Options include "LCZ" and "manual".}

\item{Turban}{Urban station references, if the method "manual" is selected.}

\item{Trural}{Rural station references, if the method "manual" is selected.}

\item{by}{The time-series split option. Options include \dQuote{year}, \dQuote{season}, \dQuote{seasonyear}, \dQuote{month}, \dQuote{monthyear}, \dQuote{weekday}, \dQuote{weekend}, \dQuote{site}, \dQuote{daylight}, or \dQuote{dst} (daylight saving time). See the \code{type} argument in the \code{openair} package: \url{https://bookdown.org/david_carslaw/openair/sections/intro/openair-package.html#the-type-option}.}

\item{impute}{Method to impute missing values. Options include \dQuote{mean}, \dQuote{median}, \dQuote{knn}, or \dQuote{bag}.}

\item{group}{If TRUE urban, rural and UHI - related temperatures are grouped in the same plot.}

\item{iplot}{Set to \code{TRUE} to return a plot. If \code{FALSE}, a data frame is returned.}

\item{isave}{Set to \code{TRUE} to save all results (plot, time-series, station references) into your directory.}

\item{ylab}{The label for the y-axis.}

\item{xlab}{The label for the x-axis. Default is \dQuote{Time}.}

\item{ylab2}{The label for the right y-axis.}

\item{title}{The title of the plot. Default is an empty string.}

\item{caption}{The source data caption.}
}
\value{
A visual representation of the time series of UHI in \code{ggplot} or data frame .csv format.
}
\description{
This function calculates the Urban Heat Island (UHI) intensity based on air temperature measurements and Local Climate Zones (LCZ). More details: \url{https://bymaxanjos.github.io/LCZ4r/articles/Introd_local_LCZ4r.html}
}
\examples{
\dontrun{
# Calculate hourly UHI intensity for 2019 using the LCZ method
my_uhi <- lcz_uhi_intensity(lcz_map, data_frame = lcz_data, var = "airT",
                            station_id = "station", year = 2019,
                            time.freq = "hour", method = "LCZ")


}

}
\seealso{
See the documentation for \code{lcz_get_map()} to obtain an LCZ map.
}
\keyword{Climate}
\keyword{LCZ,}
\keyword{Local}
\keyword{Zone,}
\keyword{analysis}
\keyword{climate,}
\keyword{spatial}
\keyword{urban}
