% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCZarea.R
\name{LCZarea}
\alias{LCZarea}
\title{Calculate LCZ class areas}
\usage{
LCZarea(lcz_map, iplot = TRUE)
}
\arguments{
\item{iplot}{}
}
\value{
areas em percentage and km2
}
\description{
Calculate LCZ class areas
}
\examples{
myLCZarea <- LCZarea(map, iplot = TRUE)

}
