% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCZarea.R
\name{LCZarea}
\alias{LCZarea}
\title{Calculate LCZ Class Areas}
\usage{
LCZarea(x, iplot = TRUE, isave = TRUE)
}
\arguments{
\item{x}{A raster SpatRaster layer containing LCZ classes.}

\item{iplot}{Logical, indicating whether to create a plot (default is TRUE).}

\item{isave}{Save the plot into your directory.}
}
\value{
A summary table of LCZ class areas if iplot is FALSE, otherwise, a bar plot.
}
\description{
This function calculates the areas of LCZ (Local Climate Zones) classes in both percentage and square kilometers.
}
\examples{

myLCZarea <- LCZarea(lcz_map, iplot = TRUE, isave=TRUE)

}
\seealso{
See the documentation for getLCZmap() to obtain an LCZ map.
}
